Changes to Testbase class
