--- orangehrm-batch26/src/test/java/ca/qaguru/orangehrmbatch26/tests/ParameterizationTestData.java
@@ -1,14 +1,0 @@
-package ca.qaguru.orangehrmbatch26.tests;
-
-import org.testng.annotations.DataProvider;
-
-public class ParameterizationTestData {
-    @DataProvider(name = "LoginInfo")
-    public Object[][] getUserInfo(){
-        return new Object[][]{
-                {"admin","admin123"},
-                {"joe", "joe123"},
-                {"Hary", "Password"}
-        };
-    }
-}
